+++ src/Lakona.Game.Server.Hotfix.Abstractions/Properties/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Lakona.Game.Server.Hotfix.Tests")]
-[assembly: InternalsVisibleTo("Lakona.Game.Server.Tests")]