+++ src/Lakona.Game.Server/Hotfix/Runtime/Loading/HotfixAssemblyLoadContext.cs
@@ -1,6 +1,5 @@
 using System.Reflection;
 using System.Runtime.Loader;
-using Lakona.Game.Server.Hotfix.Abstractions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -8,7 +7,6 @@ namespace Lakona.Game.Server.Hotfix.Loading;
 
 internal sealed class HotfixAssemblyLoadContext : AssemblyLoadContext
 {
-    private static readonly string AbstractionsAssemblyName = typeof(HotfixBehaviorOfAttribute).Assembly.GetName().Name!;
     private static readonly string RuntimeAssemblyName = typeof(IGameSessionLifecycle).Assembly.GetName().Name!;
     private static readonly string DependencyInjectionAbstractionsAssemblyName = typeof(IServiceCollection).Assembly.GetName().Name!;
     private static readonly string LoggingAbstractionsAssemblyName = typeof(ILogger).Assembly.GetName().Name!;
@@ -71,7 +69,6 @@ internal sealed class HotfixAssemblyLoadContext : AssemblyLoadContext
 
         var names = new HashSet<string>(StringComparer.Ordinal)
         {
-            AbstractionsAssemblyName,
             RuntimeAssemblyName,
             DependencyInjectionAbstractionsAssemblyName,
             LoggingAbstractionsAssemblyName